requirements.txt
tweepy
minicli
requests
jinja2
pyyaml
